Fix luna-splash graphics alignment, sync issues, and terminal restore

+++ src/luna-init/shutdown.c
@@ -35,7 +35,7 @@
 volatile int g_shutting_down = 0;
 
 static void sleep_ms(int ms) {
-    struct timespec ts = { .tv_sec = ms / 1000, .tv_nsec = (ms % 1000) * 1000000 };
+    struct timespec ts = { .tv_sec = ms / 1000, .tv_nsec = (long)(ms % 1000) * 1000000L };
     nanosleep(&ts, NULL);
 }
